move parameter from device to log interface
