+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/ingestion/models/Device.java
@@ -46,8 +46,6 @@ public class Device extends WrapperSdk {
 
     private static final String CARRIER_COUNTRY = "carrierCountry";
 
-    private static final String DATA_RESIDENCY_REGION = "dataResidencyRegion";
-
     private static final String APP_BUILD = "appBuild";
 
     private static final String APP_NAMESPACE = "appNamespace";
@@ -122,11 +120,6 @@ public class Device extends WrapperSdk {
      * Carrier country code (for mobile devices).
      */
     private String carrierCountry;
-
-    /**
-     * Carrier data residency region.
-     */
-    private @Nullable String dataResidencyRegion;
 
     /**
      * The app's build number, e.g. 42.
@@ -392,25 +385,6 @@ public class Device extends WrapperSdk {
     }
 
     /**
-     * Get the dataResidency value.
-     *
-     * @return the dataResidency value
-     */
-    @Nullable
-    public String getDataResidencyRegion() {
-        return this.dataResidencyRegion;
-    }
-
-    /**
-     * Set the dataResidency value.
-     *
-     * @param dataResidencyRegion the dataResidency value to set
-     */
-    public void setDataResidencyRegion(@Nullable String dataResidencyRegion) {
-        this.dataResidencyRegion = dataResidencyRegion;
-    }
-
-    /**
      * Get the appBuild value.
      *
      * @return the appBuild value
@@ -463,7 +437,6 @@ public class Device extends WrapperSdk {
         setAppVersion(object.getString(APP_VERSION));
         setCarrierName(object.optString(CARRIER_NAME, null));
         setCarrierCountry(object.optString(CARRIER_COUNTRY, null));
-        setDataResidencyRegion(object.optString(DATA_RESIDENCY_REGION, null));
         setAppBuild(object.getString(APP_BUILD));
         setAppNamespace(object.optString(APP_NAMESPACE, null));
     }
@@ -485,7 +458,6 @@ public class Device extends WrapperSdk {
         writer.key(APP_VERSION).value(getAppVersion());
         JSONUtils.write(writer, CARRIER_NAME, getCarrierName());
         JSONUtils.write(writer, CARRIER_COUNTRY, getCarrierCountry());
-        JSONUtils.write(writer, DATA_RESIDENCY_REGION, getDataResidencyRegion());
         writer.key(APP_BUILD).value(getAppBuild());
         JSONUtils.write(writer, APP_NAMESPACE, getAppNamespace());
     }
@@ -543,9 +515,6 @@ public class Device extends WrapperSdk {
             return false;
         }
         if (carrierCountry != null ? !carrierCountry.equals(device.carrierCountry) : device.carrierCountry != null) {
-            return false;
-        }
-        if (dataResidencyRegion != null ? !dataResidencyRegion.equals(device.dataResidencyRegion) : device.dataResidencyRegion != null) {
             return false;
         }
         if (appBuild != null ? !appBuild.equals(device.appBuild) : device.appBuild != null) {
