+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/DeviceInfoHelper.java
@@ -49,11 +49,6 @@ public class DeviceInfoHelper {
      */
     private static String mCountryCode;
 
-    /**
-     * Residency region.
-     */
-    private static @Nullable String mDataResidencyRegion;
-
     public static PackageInfo getPackageInfo(Context context) {
         try {
             PackageManager packageManager = context.getPackageManager();
@@ -104,9 +99,6 @@ public class DeviceInfoHelper {
         if (mCountryCode != null) {
             device.setCarrierCountry(mCountryCode);
         }
-
-        /* Set residency region. */
-        device.setDataResidencyRegion(mDataResidencyRegion);
 
         /* Locale. */
         device.setLocale(Locale.getDefault().toString());
@@ -219,15 +211,6 @@ public class DeviceInfoHelper {
         }
         mCountryCode = countryCode;
         AppCenterLog.debug(AppCenterLog.LOG_TAG, String.format("Set country code: %s", countryCode));
-    }
-
-    /**
-     * Set the country code or any other string to identify residency region.
-     *
-     * @param dataResidencyRegion residency region code.
-     */
-    public static void setDataResidencyRegion(@Nullable String dataResidencyRegion) {
-        mDataResidencyRegion = dataResidencyRegion;
     }
 
     /**
